First attempt at changing the classes to traits
